backend/requirements.txt
# ========================================
# requirements.txt - TravelSpeek Backend
# ========================================

# FastAPI & serveur
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Base de données
sqlalchemy==2.0.35

# Authentification & sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Email
aiosmtplib==3.0.1
jinja2==3.1.4

# Traduction texte
deep-translator==1.11.4

# Traduction vocale (Whisper)
faster-whisper==1.1.0
gTTS==2.5.4
pydub==0.25.1

# OCR image
easyocr==1.7.2
Pillow==10.4.0
opencv-python-headless==4.10.0.84
torch==2.5.1
torchvision==0.20.1

# Utilitaires
numpy>=1.24.0
better-profanity==0.7.0
python-dotenv==1.0.1
httpx==0.27.2
requests==2.32.3
aiofiles==24.1.0